pyproject.toml
[project]
name = "simplelangchainapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.13.3",
    "faiss-cpu==1.10.0",
    "langchain==1.0.0",
    "langchain-community==0.3.29",
    "langchain-openai==1.1.9",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langfuse==2.60.1",
    "langgraph==1.0.8",
    "langgraph-checkpoint==4.0.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-checkpoint-sqlite==3.0.3",
    "psycopg[binary]>=3.3.3",
    "python-dotenv==1.0.1",
]
